Fix issues in markup5ever
- After enabling ALL tests there were some issues
- Also using `$ crate` instead `::markup5ever`
- Removed unused crate
