+++ markup5ever/Cargo.toml
@@ -25,5 +25,4 @@ heapsize_derive = { version = "0.1", optional = true }
 string_cache_codegen = "0.4"
 phf_codegen = "0.7.3"
 rustc-serialize = "0.3.15"
-syn = { version = "0.11", features = ["full", "visit"] }
